requirements.txt
freetype-py==2.5.1
pillow==11.3.0
PyYAML==6.0.2
